Add files via upload
The proposed methodology for building the chatbot involves several steps. First, the scope and requirements of the chatbot are defined by analysing the needs of the institution and the user base. Next, the chatbot architecture is designed as a web-based application using Python and appropriate libraries and frameworks for the backend and frontend components. The architecture includes the use of natural language processing (NLP) techniques and integration with other systems.  Then, the chatbot is developed using Python and appropriate libraries and frameworks. Dialogflow is used to provide natural language understanding and response generation capabilities, and SQL Alchemy is used as the ORM library for managing the database interactions. Finally, the chatbot is deployed in an institutional setting and monitored for performance and usage over time. 
The minimum software and hardware operating platforms are: 
Hardware requirements A server: The chatbot would require a server to host the chatbot software and database, allowing it to process and store user queries and responses. Internet connectivity: The chatbot would require a stable and reliable internet connection to access the college's databases and other online resources.  Storage: The server would require sufficient storage to store the chatbot's database, which would include placement statistics, job openings, and other relevant information.  Processing power: The server would require adequate processing power to handle multiple user queries and responses simultaneously, ensuring that the chatbot can provide quick and accurate responses to users.  Security measures: The chatbot's server would require security measures such as firewalls, encryption, and access controls to protect sensitive user data and prevent unauthorized access. Backup and disaster recovery: The server would require backup and disaster recovery measures to ensure that data is not lost in case of a hardware failure or other unexpected event.